requirements.pipeline.txt
prefect==2.14.10
griffe==0.36.5

pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
pyarrow==14.0.1

minio==7.1.16
psycopg2-binary==2.9.9
mlflow
